Audino9012Drive: Interfacing with Passive Buzzer
Adding information about the two current projects in this folder.
Added source code that interfaces the Arduino Nano with a passive
buzzer. There are two types of sound that is produced by the buzzer.
The first is a one second tick. The second in more of an alarm
buzzer sound.

	new file:   buzzer/buzzer.ino